Create k8s-infra-live.sts.yml
Could use some eyes on this as this is my first time. I am setting up workflows that will use the octo-sts/action to generate my github tokens. The workflow lives in the k8s-platform-v3-infra-live repo, and the workflow will need to generate a token that allows read permissions to the infra-modules and monorepo. Hopefully I set this up correct but I'm unsure on the syntax for the subject regarding branches.